refactor(state): inline PageHudAnim into state.rs, remove separate page_hud.rs
The struct has no deps of its own and is small enough to live inline.
HUD drawing code already lives in draw_page/hud.rs.
