+++ gui/state.rs
@@ -7,7 +7,6 @@ mod hover;
 mod multi_selection;
 mod new_project_dialog;
 mod page_cache;
-mod page_hud;
 mod pool;
 mod selection;
 mod selections;
@@ -25,7 +24,6 @@ pub use hover::HoveredTarget;
 pub use multi_selection::MultiSelection;
 pub use new_project_dialog::NewProjectDialogState;
 pub use page_cache::PageCache;
-pub use page_hud::PageHudAnim;
 pub use pool::PhotoSelection;
 pub use selection::SlotSelection;
 pub use selections::Selections;
@@ -54,6 +52,47 @@ pub struct DataState {
     pub projects: Vec<ProjectInfo>,
     /// Recent commit history for the current branch.
     pub history: Vec<HistoryEntry>,
+}
+
+pub struct PageHudAnim {
+    pub opacity: f32,
+    pub pill_width: f32,
+    pub actions_alpha: f32,
+    pub actions_offset: f32,
+}
+
+impl Default for PageHudAnim {
+    fn default() -> Self {
+        Self {
+            opacity: 0.55,
+            pill_width: 10.0,
+            actions_alpha: 0.0,
+            actions_offset: -4.0,
+        }
+    }
+}
+
+impl PageHudAnim {
+    pub fn advance(&mut self, hovered: bool, dt: f32) -> bool {
+        let k = 1.0 - (-dt / 0.15).exp();
+        let (target_opacity, target_pill, target_actions_alpha, target_actions_offset) = if hovered
+        {
+            (1.0_f32, 80.0_f32, 1.0_f32, 0.0_f32)
+        } else {
+            (0.55_f32, 10.0_f32, 0.0_f32, -4.0_f32)
+        };
+
+        let lerp = |cur: f32, tgt: f32| cur + (tgt - cur) * k;
+        self.opacity = lerp(self.opacity, target_opacity);
+        self.pill_width = lerp(self.pill_width, target_pill);
+        self.actions_alpha = lerp(self.actions_alpha, target_actions_alpha);
+        self.actions_offset = lerp(self.actions_offset, target_actions_offset);
+
+        (self.opacity - target_opacity).abs() > 0.002
+            || (self.pill_width - target_pill).abs() > 0.3
+            || (self.actions_alpha - target_actions_alpha).abs() > 0.002
+            || (self.actions_offset - target_actions_offset).abs() > 0.1
+    }
 }
 
 pub struct InteractionState {
